feat: identidad multi-proveedor GAFI -- codigo/Alterno/EAN, remision, descuento por tramo, imagenes (Fases 1-9)
Auditoria real de GAFI (9,582 productos, catalogo real de Ferreteria
Olimpico) mostro que el codigo impreso en el paquete (Alterno del
fabricante) nunca se guardaba ni quedaba buscable -- solo el codigo
interno de GAFI. Plan de 9 fases, todas implementadas y probadas:

- Fase 1-2: catalogo_productos gana clave_proveedor; catalogo_maestro_
  identificadores ahora registra codigo/fabricante/EAN sin nunca
  reasignar uno ya tomado (conflicto -> revision humana, nunca fusion
  automatica); 35 conflictos reales de GAFI-vs-TRUPER revisados a mano
  y fusionados, 1 dejado para revision fisica.
- Fase 3: aprender EAN por escaneo -- confirmar un producto ya
  identificado con su codigo de barras real lo deja buscable para
  cualquier negocio, sin pantalla nueva (se engancha en agregar/editar
  producto).
- Fase 4: Explorar Nexo ya resolvia por codigo exacto pero perdia el
  codigo de fabricante antes de llegar a pantalla, y un bug real de
  mayusculas/minusculas rompia el match de Alternos con letras.
- Fase 5: recepcion por remision (foto -> IA de vision -> mismo
  pipeline de revision que un CFDI) para no frenar mercancia; cuando
  llega la factura real, Nexo pregunta si concilia con la remision
  antes de aplicar nada, nunca duplica el stock.
- Fase 6-7: reglas de precio ganan tramos de descuento por monto de
  factura, exclusivos del proveedor que el dueño configure (nunca
  automatico para otros) -- costo neto correcto al confirmar.
- Fase 8: Gmail hereda automaticamente conciliacion y descuento (mismo
  pipeline compartido, verificado con prueba explicita).
- Fase 9: prioridad de fuentes para la imagen del Catalogo Maestro
  (antes nunca se actualizaba al enlazar un producto existente) +
  tablero de cobertura.

Bug real encontrado de paso (preexistente, no de esta fase): NULLIF($n, 0)
sin tipo hacia que Postgres asumiera entero -- cualquier costo con
centavos tronaba al confirmar una recepcion, incluida la ruta vieja de
Recepcion de mercancia. Corregido en los 3 lugares.

217 pruebas automatizadas nuevas/re-verificadas, todas contra el
servidor y la base de datos reales con negocios sinteticos.

+++ fase4-server.js
@@ -383,11 +383,20 @@ module.exports = (app, pool, normalizarCodigo, requerirAccesoNegocio) => {
                 const itemId = item.pedidoItemId || item.pedido_item_id || null;
 
                 if (productoId && cantidad > 0) {
+                    // NULLIF($2, 0) con el 0 sin tipo: Postgres decide el
+                    // tipo de $2 localmente dentro del NULLIF (entero, por
+                    // el literal), antes de que el COALESCE de afuera vea
+                    // que en realidad va a una columna numeric -- un costo
+                    // con centavos (la inmensa mayoria de los reales) tronaba
+                    // "invalid input syntax for type integer". Encontrado
+                    // real via Fase 7 (descuento por tramo, que sí produce
+                    // decimales) pero el bug ya afectaba cualquier costo con
+                    // centavos desde antes, en produccion.
                     await client.query(`
                         UPDATE public.productos
                         SET
                             stock = stock + $1,
-                            precio_distribuidor = COALESCE(NULLIF($2, 0), precio_distribuidor)
+                            precio_distribuidor = COALESCE(NULLIF($2, 0::numeric), precio_distribuidor)
                         WHERE id = $3
                         AND negocio_id = $4
                     `, [cantidad, costo, productoId, negocio.id]);
